@lucasfari-a Dashboard finalizada. Lançar valores finalizado. HomeUsu…

+++ app/src/main/java/com/example/na_regua_app/data/model/Dado.kt
@@ -3,7 +3,7 @@ package com.example.na_regua_app.data.model
 import androidx.compose.ui.graphics.Color
 
 data class Dado(
-    var value: Float,
+    var value: Double,
     var label: String,
     var color: Color? = null,
 )
